graph viz refactoring + improvements
