+++ src/nnsight/intervention/protocols/grad.py
@@ -1,4 +1,3 @@
-from collections import defaultdict
 from typing import TYPE_CHECKING, Any, Dict
 
 import torch
@@ -65,12 +64,10 @@ class GradProtocol(Protocol):
             - Dict: dictionary style.
         """
 
-        return {
-            "node": {"color": "green4", "shape": "box"},  # Node display
-            "label": cls.__name__,
-            "arg": defaultdict(
-                lambda: {"color": "gray", "shape": "box"}
-            ),  # Non-node argument display
-            "arg_kname": defaultdict(lambda: None),  # Argument label key word
-            "edge": defaultdict(lambda: {"style": "solid"}), # Argument edge display
-        }
+        default_style = super().style()
+
+        default_style["node"] = {"color": "green4", "shape": "box"}
+    
+        return default_style
+
+
