+++ src/nnsight/intervention/protocols/lock.py
@@ -1,6 +1,4 @@
-from typing import TYPE_CHECKING, Dict, Any
-
-from collections import defaultdict
+from typing import TYPE_CHECKING, Any, Dict
 
 from ...tracing.protocols import Protocol
 
@@ -31,13 +29,3 @@ class LockProtocol(Protocol):
         default_style["node"] = {"color": "brown", "shape": "ellipse"}
 
         return default_style
-
-        """ return {
-            "node": {"color": "brown", "shape": "ellipse"},  # Node display
-            "label": cls.__name__,
-            "arg": defaultdict(
-                lambda: {"color": "gray", "shape": "box"}
-            ),  # Non-node argument display
-            "arg_kname": defaultdict(lambda: None),  # Argument lable key word
-            "edge": defaultdict(lambda: {"style": "solid"}), # Argument edge display
-        } """
