+++ src/nnsight/intervention/protocols/module.py
@@ -1,4 +1,3 @@
-from collections import defaultdict
 from typing import TYPE_CHECKING, Any, Dict
 
 import torch
@@ -104,16 +103,8 @@ class ApplyModuleProtocol(Protocol):
             - Dict: dictionary style.
         """
 
-        return {
-            "node": {
-                "color": "blue",
-                "shape": "polygon",
-                "sides": 6,
-            },  # Node display
-            "label": cls.__name__,
-            "arg": defaultdict(
-                lambda: {"color": "gray", "shape": "box"}
-            ),  # Non-node argument display
-            "arg_kname": defaultdict(lambda: None),  # Argument label word
-            "edge": defaultdict(lambda: {"style": "solid"}), # Argument edge display
-        }
+        default_style = super().style()
+
+        default_style["node"] = {"color": "green4", "shape": "polygon", "sides": 6}
+
+        return default_style
