+++ src/nnsight/tracing/graph/graph.py
@@ -1,13 +1,7 @@
 from __future__ import annotations
 
-import tempfile
-from collections import defaultdict
-from collections.abc import Iterable
-from typing import (Any, Callable, Dict, Generic, Iterator, List, Optional,
-                    Set, Type, TypeVar, Union, overload)
-
-import torch
-from PIL import Image as PILImage
+from typing import (Callable, Generic, Iterator, List, Optional, Type, TypeVar,
+                    Union, overload)
 
 from ... import util
 from .. import protocols
@@ -115,188 +109,6 @@ class Graph(Generic[NodeType, ProxyType]):
             memo[node.index] = new_node.index
 
         return new_graph
-    
-    ######## Graph Visualization ###########################
-    
-    def visualize(self,
-        title: str = "graph",
-        path: str = ".",
-        display: bool = True,
-        save: bool = False,
-        recursive: bool = False,
-        group: bool = False,
-    ):
-        
-        try:
-
-            import pygraphviz as pgv
-
-        except Exception as e:
-
-            raise type(e)(
-                "Visualization of the Graph requires `pygraphviz` which requires `graphviz` to be installed on your machine."
-            ) from e
-        
-        if group and not recursive:
-            print("Warning: set `recursive=True` to visualize all subgraphs and make use of the 'group' functionality.")
-            group = False
-        
-        from IPython.display import Image
-        from IPython.display import display as IDisplay
-
-        graph: pgv.AGraph = pgv.AGraph(strict=True, directed=True)
-
-        graph.graph_attr.update(
-            label=title, fontsize="20", labelloc="t", labeljust="c"
-        )
-
-        def style_node(node):
-            """
-            
-            """
-
-            if isinstance(node.target, type) and issubclass(node.target, protocols.Protocol):
-                return node.target.style()
-            else:
-                return {
-                    "node": {"color": "black", "shape": "ellipse"},
-                    "label": (node.target if isinstance(node.target, str) else node.target.__name__),
-                    "arg": defaultdict(lambda: {"color": "gray", "shape": "box"}),
-                    "arg_kname": defaultdict(lambda: None),
-                    "edge": defaultdict(lambda: {"style": "solid"}),
-                }
-        
-        subgraphs = {}
-        subgraph_names: Dict[str, int] = defaultdict(lambda: 0)
-        def get_subgraph(node):
-            nonlocal subgraphs
-            if group:
-                if id(node.graph) != id(self):
-                    if not id(node.graph) in subgraphs.keys():
-                        subgraph = graph.subgraph(name=f"cluster_{id(node.graph)}")
-                        subgraph.graph_attr['penwidth'] = 0.25
-                        subgraphs[id(node.graph)] = subgraph
-
-                    return subgraphs[id(node.graph)]
-                else:
-                    return graph
-            else:
-                return graph
-        
-        if recursive:
-            nodes = [node for node in self.nodes if id(node.graph) >= id(self)]
-        else:
-            nodes = self
-
-        visualized_nodes = set()
-        for node in nodes:
-            styles = style_node(node)
-
-            subgraph = get_subgraph(node)
-
-            subgraph.add_node(node.index, label=styles["label"], **styles["node"])
-            visualized_nodes.add(node.index)
-
-            for idx, arg in enumerate(node.args):
-                if isinstance(arg, SubGraph):
-                    name = f"{node.index}_{arg}_{idx}"
-                    label = f"Subgraph"
-                    
-                    subgraph.add_node(name, label=label, **{"color": "purple", "shape": "box"})
-                    
-                    if recursive:
-                        for sub_node in arg:
-                            root_node: bool = True
-                            for dep_idx in sub_node._dependencies:
-                                root_node = root_node and (dep_idx not in arg.subset)
-                                
-                            if root_node:
-                                graph.add_edge(node.index, sub_node.index, **{"style": "dashed", "color": styles["node"]["color"]})
-
-                    if group:
-                        subgraph_label = styles['label']
-                        subgraphs[id(arg)].graph_attr['label'] = f"{subgraph_label}_{subgraph_names[subgraph_label]}"
-                        subgraph_names[subgraph_label] += 1
-                    
-                elif isinstance(arg, Node):
-                    name = arg.index
-                    label = node.index
-
-                    if arg.index not in visualized_nodes:
-                        arg_label = (node.target if isinstance(node.target, str) else node.target.__name__)
-
-                        subgraph.add_node(arg.index, label=arg_label, **{"color": "brown", "shape": "box"})
-
-                        visualized_nodes.add(arg.index)
-                else:
-                    name = str(node.index)
-                    if isinstance(arg, torch.Tensor):
-                        name += f"_Tensor_{idx}"
-                        label = "Tensor"
-                    elif isinstance(arg, str):
-                        name += f"_{arg}_{idx}"
-                        label = f'"{arg}"'
-                    else:
-                        name += f"_{arg}_{idx}"
-                        label = str(arg)
-
-                    if not styles["arg_kname"][idx] is None:
-                            label = f"{styles['arg_kname'][idx]}={label}"
-                    else:
-                        label = label
-
-                    subgraph.add_node(name, label=label, **{"color": "gray", "shape": "box"}) # same
-
-                    if isinstance(arg, Iterable):
-                        for idx, element in enumerate(arg):
-                            if isinstance(element, Node):
-                                if element.index not in visualized_nodes:
-                                    """ styles = (
-                                        element.target.style() 
-                                        if isinstance(element.target, type) and issubclass(element.target, protocols.Protocol) 
-                                        else default_style(element)
-                                    )
-
-                                    styles["node"]["color"] = "brown"
-                                    styles["node"]["shape"] = "box" """
-
-                                    element_label = (element.target if isinstance(element.target, str) else element.target.__name__)
-                                    subgraph.add_node(element.index, label=element_label, color="brown", shape="box")
-                                    visualized_nodes.add(element.index)
-                                graph.add_edge(element.index, name, style="dashed", color="gray", label=f"{idx}", fontsize=10)
-                    
-                subgraph.add_edge(name, node.index, **styles["edge"][idx])
-
-        def display_graph(file_name):
-            in_notebook = True
-
-            # Credit: Till Hoffmann - https://stackoverflow.com/a/22424821
-            try:
-                from IPython import get_ipython
-
-                if "IPKernelApp" not in get_ipython().config:
-                    in_notebook = False
-            except ImportError:
-                in_notebook = False
-            except AttributeError:
-                in_notebook = False
-
-            if in_notebook:
-                IDisplay(Image(filename=file_name))
-            else:
-                img = PILImage.open(file_name)
-                img.show()
-                img.close()
-
-        if not save:
-            with tempfile.NamedTemporaryFile(suffix=".png") as temp_file:
-                graph.draw(temp_file.name, prog="dot")
-                if display:
-                    display_graph(temp_file.name)
-        else:
-            graph.draw(f"{path}/{title}.png", prog="dot")
-            if display:
-                display_graph(f"{path}/{title}.png")
 
     ### Magic Methods ######################################
