+++ src/nnsight/tracing/protocols/variable.py
@@ -1,4 +1,3 @@
-from collections import defaultdict
 from typing import TYPE_CHECKING, Any, Dict
 
 from . import Protocol
@@ -28,10 +27,8 @@ class VariableProtocol(Protocol):
             - Dict: dictionary style.
         """
 
-        return {
-            "node": {"color": "blue", "shape": "box"}, # Node display
-            "label": cls.__name__,
-            "arg": defaultdict(lambda: {"color": "gray", "shape": "box"}), # Non-node argument  
-            "arg_kname": defaultdict(lambda: None), # Argument label key word
-            "edge": defaultdict(lambda: {"style": "solid"}) # Argument edge display
-        }
+        default_style = super().style()
+
+        default_style["node"] = {"color": "blue", "shape": "box"}
+        
+        return default_style
